Move some functions to gitutil/gitutil.go
